Conflicts:
	include/pattern.h

+++ src/benchmarks.cpp
@@ -175,11 +175,11 @@ void naivetrain(string filename, std::unordered_map<string,uint32_t> & model, in
           getline(IN, line);
           vector<string> tokens;
           split(line, ' ', tokens);
-          for (int i = 0; i < tokens.size(); i++) {
+          for ( size_t i = 0; i < tokens.size(); i++) {
             string token = tokens[i];
             for (int n = 1; n <= maxlength && i+n <= tokens.size(); n++) {
-                  if (n > 1) token += " " + tokens[i+n-1];
-                  model[token] += 1;
+	      if (n > 1) token += " " + tokens[i+n-1];
+	      model[token] += 1;
             }
           }
        }
@@ -352,7 +352,7 @@ int main( int argc, char *argv[] ) {
     }
 
 
-    
+
     if (testnr == 99) {
         Measurement m = begin(string("99 - Training indexed PatternModel from preloaded corpus: threshold 2, up to 8-grams, with skipgrams, skip-type threshold 12"));
         PatternModelOptions options;
@@ -404,4 +404,3 @@ int main( int argc, char *argv[] ) {
         end(m);
     }*/
 }
-
